requirements.txt
Fabric3==1.12.post1
python-dotenv==0.5.1
pyyaml==3.11
